03_submission/requirements.txt
pillow
SimpleITK
imagecodecs
numpy
opencv-python-headless
timm